Move ComponentList and Tools into kapture-backend
