+++ ui/src/implementations/kapture-backend/components/editor/Tools/ApplyTool/ApplyTool.js
@@ -7,9 +7,11 @@ import { CompoundsSection } from './CompoundsSection';
 import { MediaSection } from './MediaSection';
 import { SupplementsSection } from './SupplementsSection';
 import { AttributesSection } from './AttributesSection';
-import { groupComponents } from '../../../../store/plateFunctions';
+import { plateFunctions } from '../../../../store';
 import { SelectedContainers } from './SelectedContainers';
 import styles from './ApplyTool.module.css';
+
+const { groupComponents } = plateFunctions;
 
 export class ApplyTool extends Component {
   groupComponents = memoize(groupComponents);
