+++ ui/src/implementations/kapture-backend/components/editor/Tools/ApplyTool/ComponentsSection.js
