+++ ui/src/implementations/kapture-backend/components/editor/Tools/ApplyTool/CompoundsSection.js
