+++ ui/src/implementations/kapture-backend/components/editor/Tools/ApplyTool/SupplementsSection.js
