+++ ui/src/implementations/kapture-backend/components/editor/Tools/ApplyTool/ToolComponent/ToolComponent.js
