+++ ui/src/implementations/kapture-backend/components/editor/Tools/ClearTool/ClearTool.module.css
