+++ ui/src/implementations/kapture-backend/components/editor/Tools/ClearTool/connect.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 
 import { ClearTool } from './ClearTool';
 import { selectors } from 'AtlasUI/store';
-import { actions } from '../../../../implementations/kapture-backend/actions/';
+import { actions } from '../../../../actions';
 
 const { selectEditorClearMode } = selectors;
 const { setClearMode } = actions.editorTools;
