+++ ui/src/implementations/kapture-backend/components/editor/Tools/ClearTool/index.js
