+++ ui/src/implementations/kapture-backend/components/editor/index.js
@@ -1,0 +1,1 @@
+export { EditorSidebar } from './EditorSidebar';
